+++ packages/components/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.1.0...@ant-design/pro-components@2.1.1) (2022-08-25)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+# [2.1.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@1.1.19...@ant-design/pro-components@2.1.0) (2022-08-25)
+
+### Features
+
+- **layout:** layout support simple app list ([#5501](https://github.com/ant-design/pro-components/issues/5501)) ([2fe76ed](https://github.com/ant-design/pro-components/commit/2fe76edea9acd07d1f167762ebd52f6fe7a163a2))
+
 ## [2.0.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.0.0-experimental.30...@ant-design/pro-components@2.0.5) (2022-08-24)
 
 **Note:** Version bump only for package @ant-design/pro-components